Track file paths instead of file hashes
Renamed seen_files to extracted_files.
Switched to storing the file path so that extracted_files can be
used as the set of extracted file paths to add to a request.

The returned files is now a set and now includes all files in the tree,
even if they have been seen before for consistent output.
If the file has been seen before writting the file is skipped as
it should already have been written.